+++ JotunnLib/Utils/PieceUtils.cs
@@ -1,22 +1,22 @@
 using System;
-using System.Linq;
 
 namespace Jotunn.Utils
 {
     internal static class PieceUtils
     {
-        public static int VanillaMaxPieceCategory { get; } = GetVanillaMaxPieceCategory();
+        public static Piece.PieceCategory VanillaMaxPieceCategory { get; } = GetVanillaPieceCategory(nameof(Piece.PieceCategory.Max), Piece.PieceCategory.Max);
+        public static Piece.PieceCategory VanillaAllPieceCategory { get; } = GetVanillaPieceCategory(nameof(Piece.PieceCategory.All), Piece.PieceCategory.All);
 
-        private static int GetVanillaMaxPieceCategory()
+        private static Piece.PieceCategory GetVanillaPieceCategory(string name, Piece.PieceCategory fallback)
         {
             try
             {
-                return (int)Enum.Parse(typeof(Piece.PieceCategory), nameof(Piece.PieceCategory.Max));
+                return (Piece.PieceCategory)Enum.Parse(typeof(Piece.PieceCategory), name);
             }
             catch (Exception e)
             {
-                Logger.LogWarning("Could not find Piece.PieceCategory.Max, using fallback value 4");
-                return 4;
+                Logger.LogWarning($"Could not find Piece.PieceCategory {name}, using fallback value {fallback}");
+                return fallback;
             }
         }
     }